backend-ai/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
transformers>=4.50.0
torch>=2.0.0
underthesea>=6.8.0
deep-translator>=1.11.4
langdetect>=1.0.9
emoji>=2.14.0
pymongo[srv]>=4.7.0
python-dotenv>=1.0.1
numpy>=1.24.0
